move tests to generated-file.test.ts
