+++ packages/protoplugin-test/src/transpile.test.ts
@@ -202,75 +202,4 @@ describe("transpile", function () {
       "export declare function foo(): Foo;",
     ]);
   });
-
-  describe("print with tagged template literal", function () {
-    test("one line with symbol", () => {
-      const linesOf = transpile((schema) => {
-        const f = schema.generateFile("test.ts");
-        const Foo = f.import("Foo", "foo");
-        f.print`export function foo(): ${Foo} { return new ${Foo}(); };`;
-      });
-      expect(linesOf("test.ts")).toStrictEqual([
-        'import {Foo} from "foo";',
-        "",
-        "export function foo(): Foo { return new Foo(); };",
-      ]);
-    });
-    test("multi lines with symbol", () => {
-      const linesOf = transpile((schema) => {
-        const f = schema.generateFile("test.ts");
-        const Foo = f.import("Foo", "foo");
-        f.print`export function foo(): ${Foo} {
-  return new ${Foo}();
-};`;
-      });
-      expect(linesOf("test.ts")).toStrictEqual([
-        'import {Foo} from "foo";',
-        "",
-        "export function foo(): Foo {",
-        "  return new Foo();",
-        "};",
-      ]);
-    });
-    test("with empty lines", () => {
-      const linesOf = transpile((schema) => {
-        const f = schema.generateFile("test.ts");
-        const Foo = f.import("Foo", "foo");
-        f.print`
-export function foo(): ${Foo} {
-
-  return new ${Foo}();
-};
-`;
-      });
-      expect(linesOf("test.ts")).toStrictEqual([
-        'import {Foo} from "foo";',
-        "",
-        "",
-        "export function foo(): Foo {",
-        "",
-        "  return new Foo();",
-        "};",
-      ]);
-    });
-    test("empty literal", () => {
-      const linesOf = transpile((schema) => {
-        const f = schema.generateFile("test.ts");
-        f.print``;
-      });
-      expect(linesOf("test.ts")).toStrictEqual([""]);
-    });
-    test("with only symbol", () => {
-      const linesOf = transpile((schema) => {
-        const f = schema.generateFile("test.ts");
-        const Foo = f.import("Foo", "foo");
-        f.print`${Foo}`;
-      });
-      expect(linesOf("test.ts")).toStrictEqual([
-        'import {Foo} from "foo";',
-        "",
-        "Foo",
-      ]);
-    });
-  });
 });
